fix(core): share benchmark timestamp parsing
