+++ packages/bench/src/adapters/remnic-adapter.ts
@@ -16,6 +16,7 @@ import {
   expandTildePath,
   normalizeTurnExpansionEnd,
   Orchestrator,
+  parseFlexibleIsoTimestamp,
   parseConfig,
 } from "@remnic/core";
 import type {
@@ -373,44 +374,9 @@ function normalizeDrainTimeoutMs(value: unknown): number {
   return value;
 }
 
-function benchDaysInMonth(year: number, month: number): number {
-  const leap = year % 4 === 0 && (year % 100 !== 0 || year % 400 === 0);
-  return [31, leap ? 29 : 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31][month - 1] ?? 0;
-}
-
 function parseStrictBenchTimestamp(value: string, label: string): number {
-  const match = value.match(
-    /^(\d{4})-(\d{2})-(\d{2})(?:[Tt](\d{2}):(\d{2})(?::(\d{2})(?:\.\d{1,9})?)?(?:[Zz]|([+-])(\d{2}):?(\d{2}))?)?$/,
-  );
-  if (!match) {
-    throw new Error(`${label} must be a valid timestamp; received ${value}`);
-  }
-
-  const year = Number(match[1]);
-  const month = Number(match[2]);
-  const day = Number(match[3]);
-  const hour = match[4] === undefined ? 0 : Number(match[4]);
-  const minute = match[5] === undefined ? 0 : Number(match[5]);
-  const second = match[6] === undefined ? 0 : Number(match[6]);
-  const offsetHour = match[8] === undefined ? 0 : Number(match[8]);
-  const offsetMinute = match[9] === undefined ? 0 : Number(match[9]);
-
-  if (
-    month < 1 ||
-    month > 12 ||
-    day < 1 ||
-    day > benchDaysInMonth(year, month) ||
-    hour > 23 ||
-    minute > 59 ||
-    second > 59 ||
-    offsetHour > 23 ||
-    offsetMinute > 59
-  ) {
-    throw new Error(`${label} must be a valid timestamp; received ${value}`);
-  }
-
-  const parsed = Date.parse(value);
-  if (!Number.isFinite(parsed)) {
+  const parsed = parseFlexibleIsoTimestamp(value);
+  if (parsed === null) {
     throw new Error(`${label} must be a valid timestamp; received ${value}`);
   }
   return parsed;
